+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_out/reset_password/FindPasswordPage.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.snutt2.views.logged_out
+package com.wafflestudio.snutt2.views.logged_out.reset_password
 
 import androidx.activity.compose.BackHandler
 import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
